// src/shared/types/types.ts
export interface ButtonProps {
  text: string;
}

export interface ChooseCardProps {
  title: string;
  status: string;
  location: string;
  type: string;
  img: string;
}

export interface BoxProps {
  text: string
}

export interface RoyalSize {
  width: number;
  height: number;
}

export interface RoyalPosition {
  top: string;
  left: string;
  width: number;
  height: number;
}